Inline EntityParserOutputDataUpdater into EntityParserOutputGenerator
As said in
https://gerrit.wikimedia.org/r/#/c/244643/3/repo/includes/DataUpdates/EntityParserOutputDataUpdater.php
and other comments the border between these two classes is not very obvious.
Both are supposed to generate and/or update the parser output. Also what we
currently call "data updater" is so simple that it's barely worth the cost
of an additional class.

This patch removes the composition aspect, but not the class. This patch is
supposed to be as simple as possible. More to come in later patches.

Bug: T114220
Change-Id: I3dfe8eca6533d37002bdfc6cbd196879d3410632