Construct DataUpdaters in EntityParserOutputGenerator instead
... instead of in the factory. This does have a important reason: The
ParserOutputDataUpdater interface and all classes implementing this
interface should be changed so they do not get the ParserOutput via
the final "updateParserOutput" call but in the constructor instead.
This allows for much more flexible and less resource intense updaters.
Instead of collecting everything in an array, the updaters can pass
stuff to the ParserOutput whenever they want. This is currently not
possible.

Still to do:
* Find a better split for all the stuff that's currently going on in
  EntityParserOutputGenerator and EntityParserOutputDataUpdater. The
  current split and boundaries between these two classes are not the
  best we can do.
* Introduce a more general StatementMatcher interface and make use
  of it, and make the PropertyDataTypeMatcher an implementation of that.
* Find better class names for several of these classes and interfaces.

Bug: T114220
Change-Id: I1aa589038bc018ad243fbaa101a5bbcd9e844281