Clean up LanguageFallbackLabelDescriptionLookupFactory and related
Motivation to touch all this code came from two issues I found in
LanguageFallbackLabelDescriptionLookupFactory while reviewing code that
uses this service:

1. newLabelDescriptionLookup did not had a description of the $entityIds
parameter. And it was not optional but should, because prefetching is
optional.

2. The default for $fallbackMode in the constructor was -1, which was
then mapped to a combination of all modes, which is identical to the
...ALL mode.

I also made use of the ...ALL constant whenever possible.

Change-Id: I90c3c428e2872d07d5bbdf087d707c2be66e2de1