Update patch set 1
Patch Set 1:

is there any chance that the API can expose metadata to indicate that 'identifier' (and commonsMedia, etc) can be handled as 'string'.
Ideally clients can _easily_ provide limited support for datatypes which are serialised as a plain string.
Clients could derive this from the JSON, but it would be nice if there was an explicit description of the datatype serialisation, where that serialisation is 'simple'.

Patch-set: 1
Label: Verified=0