Script for populating geodata and pageimages
I am not sure refreshLinks is going to work so
well for us, and in case it is a problem then
I propose a script for our specific use case.

still would like to add at least some basic tests,
if we want/need the script. 

Also, the LinksUpdate code in core could be refactored
some to make bits of the code more re-usable such
as for updating page props.

Bug: T114868
Change-Id: I39a439df9cb34c549d90a7ad288f03f7352bbb74