Split code and fix precision edge case in MwTimeIsoFormatter
Meant as a clean up of existing code. I started this patch with
splitting the big function.

I found an issue where this formatter incorrectly formats values
with hour, minute or second precision as a day. The issue with this
is that you can not even see the hour, minute or second. It's
supressed. This patch introduces a fallback to the full timestamp.

Change-Id: If5974e808f78a5fbac82f0eeb8d825509c9495a9