Update patch set 1
Patch Set 1: Code-Review+2

In the Wikidata team we think it's really, really critical to have tests that cover such bugs to not constantly run into regressions. We would very much appreciate if you could add tests. In this case I did that in the following patch. It was not only this feature that had no test, the whole function was not tested.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2