Update patch set 2
Patch Set 2: Code-Review-1

I think the names of the variables and the classes should be changed consistently in a single patch, or left as they are. Renaming the variables but not the classes is not an improvement, it makes the code harder to understand.

Patch-set: 2
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1