Update patch set 2
Patch Set 2:

Prefer what?

Neither "January 1 ACE" nor "January 1 after Christ" is more obvious than "January 0001". More the opposite:
* What is "ACE"? https://de.wikipedia.org/wiki/ACE
* So "January 1 after Christ" means "some January the 1st after Christ"? Still the same confusion.

So we are again discussing if fixing a bug needs a decision on a "product level"? Whatever that means. What's the process to do that? There is a bug, here is a fix, but neither the bug nor the fix are "big" enough to get enough attention to be considered in a sprint, so stuff like this will never be fixed within years, even if the fix is sitting around. This makes me sad and angry. We are fixing dozens of bugs each day. I do not remember us making such a fuzz about the patches that fix them.

Patch-set: 2