Update patch set 2
Patch Set 2:

The decision to always have for digits in a year was for the internal representation, not the external one. 

This issue is not a technical bug, it's a usability issue. It is not obvious what solution is best, and programmers like me and you are often not the best people to judge this. So I recommend to talk about this with Lydia, or bring it up at Story Time (that's the procedure for usability / product question: schedule them for Story Time or talk to Lydia).

Patch-set: 2