Update patch set 2
Patch Set 2:

Hi no it doesn't. If you compare a previous patch which runs npm and compare it to this patch npm it shows that this patch lungs more js files then that one. And it doesn't seem jshintignore won't work if you have **/*. Or **/.js set. It was causing it to look into the node_modules folder.

Patch-set: 2
Reviewer: Paladox <1665@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0