Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

Fine for me, even if I think this patch should be split in two or more. One for the removal of the constant, one for the removal of all the supportsRedirects stuff which is not really bound to the experimental constant.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1