Don't make broken wikitext links for sister project links
It's better not to link sister projects in client recent changes
and watchlist, vs. making broken wikitext links.

I think it is somewhat non-trivial to add a decent mechanism
to build interwiki wikitext links for sister projects, and
this is a short term solution that at least makes things
less broken.

The ideal solution would be to have such decent mechanism
for making interwiki links for sister projects, with the
correct site group prefix + subdomain for applicable projects
and handle special cases like wikidata itself, with 'd' prefix
and no subdomain.

Also, added test case for this and while touching the tests,
changed protected -> private and factored out some of the code
that had been duplicated in numerous places in the test file.

and removed the wikitext link formatting in the
RecentChangeFactoryTest, since the SiteStore mock does
not actually know about any sites and thus the links
would be unlinked in this case.

Bug: T116459
Change-Id: Iabc4ea9a82c90ec9db899139cb7b5eed862e4cc6