Update patch set 2
Patch Set 2: Code-Review+2

(2 comments)

Note that this patch may introduce a new unexpected behavior: When a primary coordinate is already set from an other source, the coordinate considered "primary" by this code was previously set as a secondary coordinate. With this patch this coordinate is completely skipped and does not appear as a secondary coordinate any more.

I do not know if this is an issue that should be fixed. If you think so, please create a Phabricator ticket.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2