Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Er, why ...?

If the entity is an item, call a method that accepts an item.

If the entity is a statementprovider, call a method that accepts a statementprovider.

What's wrong with that?

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1