Add @covers tags for abstract base classes in repo
I'm focusing on abstract classes that are reported as "top project
risks" by PHPUnits coverage report.

I did not checked if a specific test contains test cases that are
supposed to test code from a base class. This would be possible but
is (at least at the moment) not necesarry in my opinion. I'm simply
adding the whole chain of base classes as @covers tags.

Change-Id: I63f0309b7b1d9812994c225b50f9f8650481f35f