Always mock LanguageNameLookup in tests
A call to LanguageNameLookup::getName is expensive and should never
happen in tests that are not supposed to cover this class.

Whenever possible I'm using "never()" to make sure this code is not
called.

A straight "getMock()" with no "expects()" means all methods in the
class are mocked with empty mocks and return null.

Change-Id: I6da5758bbb42401a10f019a484c2b37389784e11