Update patch set 1
Patch Set 1: Code-Review-1

(4 comments)

Early review. I'm aware this is work in progress.

Suggestion: Instead of resizing via background-size you can simply change the width/height of the SVGs.

Note that we need PNG fallbacks. But these can be added in a later patch.

Note there are also a few whitespace issues in the CSS files.

Please add a note to this patchs commit message what the source of the images is.

Patch-set: 1
Label: Code-Review=-1