Fix uncaught exception in Special:ItemDisambiguation
The exception happens with some invalid language codes,
such as "<omg>" though "omg" is considered "valid".

This makes validation of the language code happen before
doing search.  If the language code is invalid, then show
an error message with formatting that this an error.

Also, make use of ContentLanguages to decouple
the code from the Language class. WikibaseContentLanguages
handles stuff like $wgExtraLanguageNames.

As well, add validation in TermSearchInteractor and throw an
InvalidArgumentException if the code is invalid, instead of
hitting this later during handling language fallback in
some cases. This also avoids doing unnecessary database requests.
If the language code is invalid, then search would find no
terms anyway unless there is bad data in the terms table.

Bug: T118556
Change-Id: I1c1d73acff9d55d326ac6f65b3be40103ed87e56