Register factory callbacks for value types as well as data types.
This allows all kinds of factory callbacks to be registered by
value type instead of per property data type. This avoids redundant
declaration, and provides a fallback mechanism for new/unknown
data types.

In addition, this allows extensions to add handling for new value
types, not just data types.

Note that teh fallback my be applies by DataTypeDefinitions (when
using the default RESOLVED_MODE) or by factory classes that are
aware of the VT and PT prefixes (using PREFIXED_MODE).

Bug: T118499
Change-Id: I0666f1ea97a3a3caf7233f28c880452e891a21b6