ChangeDispatcher should use locks on the local DB.
Previously, we places global locks on the respective client DB.
In order to avoid opening database connections, we should indetead
manage the locks for all clients on the repo database.

Note: while this will likely solve the "many connections" part of T118162,
it does nothing to clarify the original cause of that issue.

Bug: T118162
Change-Id: Ibea2f574de6398f805cbc2276d6a93d9780e4b9f