Introduce and start using StatementSectionsView
This patch introduces a new view. This new view renders the "Statements"
heading and the statements below via the existing StatementGroupListView.
StatementGroupListView does *not* render the heading any more.

Note that there was no test for the heading before. We changed the
StatementGroupListView tests along with the new section view tests.

This patch is "incomplete" on purpose to make it small and easy to review.
We will continue working on the TODOs in the next patch.

Bug: T118949
Change-Id: Iac913191cdd980ed8f9659a3bcb075add6cdef39