Add class="mw-editsection" to headlines on entity pages
MediaWiki core's headlines are rendered like this:
<h2><span class="mw-headline" id="...">...</span><span class="mw-editsection">...</span></h2>

This addition is critical because I5c989f7 will change MobileFrontend in a
way that it only considers headlines with this class.

Warning, this may have unwanted consequences. For example, if a gadget
is looking for this class name and adds tools to each headline, these tools
will most probably not work on entity pages. But this will be an issue with
the gadget then and should not be a reason to not have these class names.

Bug: T110902
Change-Id: I1aa54a6a12540b6a8ff18ef0b070504ac28c4cbe