Remove deprecated "tipsy" tooltip CSS
The deprecated code is no longer needed. Be aware that this patch will
change all our blue tooltip windows to white. PM said yes to this change.
The red error tooltips will stay red.

The arrow was made invisible in I095d9b0. I could not find a reason why
this was done. Most probably because of the bug I'm now fixing with the
additional "top: 0". Also see I560d501.

Change-Id: Ifb0b9d8d4174a7701df021f729f48dd298d67c5c