Update patch set 5
Patch Set 5: Verified-1

V-1 to keep this from accidentally being merged and deployed with the regular train. See the deployment note in the commit message. It will probably be best to backport & SWAT this, so we can easily stop the scripts while the new code is rolled out.

Patch-set: 5
Label: Verified=-1