Strip HTML comments in TemplateRegistry
We just realized that 5% of our HTML is comments. These comments are
in our templates.php file to make it easier for team members and other
developers to understand the templates and how they are supposed to be
used. These comments are not for any user of the resulting HTML.

Bug: T119077
Change-Id: I3ce489f5695181f57df84bd5f6dda82fffd8ae58