Make Wikibase watchlist filter in client work with new checkboxes
The way that core generally handles this is to check if
action=submit is in the request, and if so then don't set
the default to whatever the user preference is.

We need to add a similar check in the hook handler.

RecentChanges doesn't (yet) have the checkboxes and
doesn't have action=submit when submitting the form,
so will just keep doing what was done before there.
If/when the checkboxes get introduced there, then
our code will already be compatible.

Also added tests and refactored the tests some to
make them more clear.

Bug: T118459
Change-Id: Ife7a52f3bea0a4a28ed7cc4bc6433709f8beb41b
(cherry picked from commit a672a5cd8ab68df74d419d8f63a2e2e738f410d7)