Update patch set 2
Patch Set 2: Code-Review-1

(9 comments)

This changes the behaviour of the normalization and comparison code for triples. This doesn't cause failures now, but it would cause failures for valid rdf mappings that use slightly different formatting.

In any case, instead of messing with the normalization code in 20 places, I suggest to factor it out into a separate helper class.

Patch-set: 2
Label: Code-Review=-1