Update patch set 15
Patch Set 15:

also, we already have a pencil for the "Add links" widget in the sidebar:

e.g. can see it here:

https://www.wikidata.org/wiki/Wikidata:Interwiki_conflicts

that pencil is more a gray color.  would be nice to have them use a unified icon and color.  that pencil is not yet svg, but we have a task for that:

https://phabricator.wikimedia.org/T67527#1820354

Patch-set: 15