Update patch set 2
Patch Set 2:

I think i added this for debugging convenience. Not sure it's used anywhere in the code base, but it's sure useful during debugging, whether you are using a debugger or print() or wfDebug(). 

I suggest to keep it if it doesn't cause problems. Without it, it's pretty tricky to see at a glance which change is which during a debug run.

Patch-set: 2
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0