Consolidate watchlist / recent changes query hook handling
Previously the watchlist and recent changes hooks were doing
somewhat different things, with the watchlist doing too much
and too complicated.

All the recent changes hook did was add a where condition
(rc_type != RC_EXTERNAL) when we want Wikibase changes hidden.
This is what we should do as well for the watchlist, and
if Wikibase changes are to be displayed then we don't need
to interfere with what core does.

This patch also fixes the issue of watchlist showing more
than the last change for a page, if the "Expand watchlist to show
all changes, not just the most recent" option is turned off
in preferences.

Some of the code and the dependencies needed for this hook
and the ChangesListSpecialPageFilters hook are the same
so they are handled here in the same hook handler class.

Bug: T64798
Bug: T50047
Change-Id: Ib660a016c075c51660b18cea777a4431537ecb69