Update patch set 1
Patch Set 1:

I disagree that this is a reason for a -1. Especially since this and the other patches I recently did does not fix *all* line length issues. I'm picking a few that are easy to fix. An actual patch that reduces the line length limit to 120 (which would be a good, reasonable compromise in my opinion) would be way, way larger and way, way harder to review.

This patch *is* an improvement. It makes it much easier to actually reduce the line length limit in an *other* patch.

Patch-set: 1