Update patch set 1
Patch Set 1:

Who is "we"? Personally I do not care much. I'm quite happy with the current 140. But I want readable code. This is not exclusively about line length. Therefore I'm not blindly splitting lines in this patch just because they are longer than a certain limit (even if this could be misread from my commit message). What this patch does is making code more readable. That's all. Yes, I picked candidates from a PHPCS log, but I did not used it as a bible but as a guide.

Patch-set: 1
Label: Verified=0