Update patch set 6
Patch Set 6:

(1 comment)

I just realized that a lot of my confusion comes from the fact that both the class level documentation and the tests talk about "data types", but the InfoStore interface is not used for that. It's exclusively used to query the "formatterURL" field from the wb_property_info table, not the type.

Please resolve this confusion. First, by removing all mentions of data types. Second, preferably by hardcoding the key and not passing it in via the constructor.

Patch-set: 6