Move statement filtering from GetClaims API to dedicated class
(This is pure refactoring. Should be 0 behaviour changes or
capability additions/removals.)

This moves the code out of the API module, imporving decoupling
and allowing for API-independent tests of its logic.

This also removes the manual filtering that was happening before
by instead using StatementList::filter. To see this simplification
on its own, have a look at
https://gerrit.wikimedia.org/r/#/c/255370/1/repo/includes/api/GetClaims.php

Change-Id: I44077aefcc2ec77be7ba489ec9ac5f264054b800