Enforce $wgMaxArticleSize for entities.
THis is implemented via a new wikibase setting, maxSerializedEntitySize,
which defaults to $wgMaxArticleSize. maxSerializedEntitySize is enforced
by EntityContentDataCodec, since that'S the only place where the serialized
size is known. The limit is applied both before deserialization, and after
serialization.

Note that the failure mode when encountering large entities is rather harsh.
This is intended as a hard limit to avoid issues on the storage layer.
A better limit mechanism should be implemented in higher level code, but
that mechanism cannot rely on size in bytes.

Bug: T119753
Change-Id: I64046eefdb9a9e5bb14959205841ab7fdcc3f0dd