Add documentation for EntityContentDataCodec defaults
This also makes it possible to omit the parameter. Before it must be
called as $codec->encode( $data, null ), which is odd. To what does
"null" refer? Is there a "null" format? Can the format be disabled
somehow? Now it can be called as $codec->encode( $data ). This should
make it more obvious that a default is used.

Change-Id: I172ccfa317f2ec0550cc106c0e22217342e3494f