Update patch set 1
Patch Set 1: Code-Review-1

(6 comments)

Can you please expand the commit message and explain this a little bit, so our English and German Wikibase developers can understand why this is relevant? In which languages is this relevant? Personally, I am aware that this is relevant in certain languages and I will merge this when the mistakes are fixed. But I would like to understand this better. Thanks in advance.

Patch-set: 1