Update patch set 15
Patch Set 15: Code-Review+1

Looks good as far as I understand the code. I agree with the intent: have generic browser tests in the repo where the code lives, and run them automaticalyl in CI. This will make it easier to write and maintain tests, and will make it obvious when we break things.

Since Jenkins sais it's runnign the tests and they pass, I assume it words as intended. If Jenkins sais the latest version is good, I'll merge it.

PS: I'm not very happy with the example test, but that should be fixed later.

Patch-set: 15