Update patch set 1
Patch Set 1: Code-Review-2

> Before it must be called as $codec->encode( $data, null ), which is odd.

This was intentional: the idea is to force the caller to explicity say what format they want, even if they want the default. With an optional parameter, it's too easy to simply forget to pass the format through. I made that experience with the optional format parameter in the Content and ContentHandler interface.

Thanks for fixing the documentation, btw. Once more, this shows that it'S a bad idea to bundle even seemingly trivial changes together in a single patch...

Patch-set: 1
Label: Code-Review=-2