Update patch set 2
Patch Set 2:

Hm... that singleton isn't needed right now, but if we move the handler for formatting changes lines here, that would potentially be called hundreds of times. Then the singleton would make sense - but then it mustn't contain anything specific to the hook invocation that caused it to be created. Currently, the hook handler singleton holds on to some of the parameters passed in via the hook.

I think a brief discussion is needed to decide how to procede here.

Patch-set: 2