Update patch set 1
Patch Set 1:

Guys, what's wrong with you? There *is* no test with an other subclass of the long deprecated Entity class. If you want to add more test cases, you are free to do so in an other patch. This patch is not blocking anybody doing anything. So what's the -1 for? Why do you think it's *my* job to fix *everything* in a single patch? This is not how this is supposed to work, I believe. I'm fixing one thing per patch, whenever possible. Sometimes I'm fixing a few tiny bits left and right, as I did here. But I will not grow patches to a "perfect" state where they become unreviewable and unmergable. I think I said that many, many times and I wonder why we are discussing this over and over again.

https://twitter.com/fioroco/status/671128541057142784

Patch-set: 1