Update patch set 1
Patch Set 1:

Ouch, even a -2. So there is not even a way forward, from your perspective? You think making code easier to read, understand and maintain is a *bad* thing?

> force the caller to explicity say what format they want, even if they want the default.

Null is neither something "explicit" nor can it be a "default". As I said, what does that mean? Is the default to have no format? Will this return empty strings then? Or does this mean that part of the serialization is skipped and you get an array instead?

If you want to pass in the default, pass in the default, but not "null". However, I do not see what this gives us. There is a syntax in PHP for defaults. That's what I do here. Nothing more, nothing less. Additionally, the default is so very much "hardcoded" anyway, that it seems really, really pointless to enforce each and every user of this class to pass in the same default again and again, when there is only one exception. This does not make things "more obvious", this is worse, because you end up with code duplication and a lot of question marks. Why is *this* format passed in here, and not the other? Is this on purpose? Is this a mistake? Why isn't it the default? And if it is the default, why is it necessary to repeat it? How can we ever change it if it's duplicated across the code base? But wait, the comment says it should never change. This raises even more question marks. Why is is a parameter then? Why is it *required* to pass in something a user of this class should most probably not touch at all?

A -2? Seriously?

> Once more, this shows that it'S a bad idea to bundle even seemingly trivial changes together in a single patch.

I'm not "bundling" anything. I'm making code more readable and easier to understand and maintain.

Patch-set: 1
Label: Verified=0