Allow null as default content format in DeferredDecodingEntityHolder
This issue came up when running ReferencedEntitiesDataUpdaterTest on a
wiki where the main namespace is an entity namespace.

MediaWikiTestCase::insertPage does not pass in a content format but
null. All the called code should expect that, as you can see in the
abstract ContentHandler::unserializeContent which explicitely says
that the $format parameter can be null.

But our EntityHandler::unserializeContent changes this method signature.
This works in PHP but is bad practice. I'm removing this default.
Instead the default from EntityContentDataCodec::getDefaultFormat should
be used.

Now the next issue comes up. DeferredDecodingEntityHolder requires the
format parameter to be a string. But it does not need to do so, because
all it does with the format is passing it to EntityContentDataCodec,
which accepts null and happily falls back to the default. So I'm
removing this limitation and now everything fits together.

Change-Id: Ie9f4edbf8e3fe2649d7731b36b12dba5b8679f57