Make ReferencedEntitiesDataUpdaterTest pass with Items in main namespace
This test fails when running on a repo where Items are in the main
namespace.

Yes, this clearly qualifies as a hack. I'm faking *some* valid JSON
content, but it is disjoint from the actual page name. This is fine for
this test, because all it cares about in the end is if the *page* exists.
The actual content is never checked. Not even the content type.

With this hack the test will succeed in all cases where the main
namespace is either wikitext or contains Items.

Please do not block this patch just because you have a better idea.
Submit your idea as a follow-up please.

Change-Id: I863ee7210d2b1ad3813558323a54ea1ccdcde388