Update patch set 1
Patch Set 1: Code-Review+2

It's a bit counter-intuitive to pass Language objects instead of only the language code. It should cause *more* code to be executed when working with objects instead of strings, shouldn't it? But no, not in these cases. I double checked all of them and can confirm the intent of this patch. The critical difference is the validation in LanguageFallbackChainFactory::newFromLanguageCode.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2