Prefer "use" over full qualified class names
I excluded \PHPUnit_Framework_TestCase, \MediaWikiTestCase and \stdClass,
and what you see in this patch is almost everything that's left. I may
have missed some (I used a regex to find these) but did not wanted to
spend to much time on this. My personal conclusion is that we should
close the ticket as resolved.

Bug: T45295
Change-Id: I908b09d8b05898f0a061d271817dc6ff3b6371e7