Add label count field to search index
For items like Q6581097 (male) and Q6581072 (female),
label count is probably also an important consideration.

although these have a lot of incoming links, so maybe
label count won't be essential. But think we should
include it and have it as an option when trying various
ways of rescoring search results.

Implementation is done in a consistent way as the other
fields we are introducing (sitelink and statement count).

Change-Id: I9b9d5eb69883412627257c9e90688af50af967b5