Fix format name strings in parsers
As far as I can tell this is not a breaking change. These constants
are not used anywhere except for exception messages. There was quite
some confusion about these constants when we introduced new parsers and
renamed existing parser classes. With this patch I'm applying the most
simple naming scheme: identical to the relevant part of the class name,
lowercase, separated with dashes.

Please merge https://github.com/DataValues/Time/pull/97 along with this.

Change-Id: Id4bfc429a5d22025f9fc0a29d7c30c7f7ffc10a9