Update patch set 1
Patch Set 1:

I see a lot of Language::factory() and $language->getCode() in this patch. I also see it removes a FIXME that's valid, in my opinion. None of these classes does need a Language object. All they need is the code and the direction.

As said, I suggest to do what the FIXME says instead.

I also wonder what motivated this patch? Does this turn out to be a bottleneck in a specific situation? If so, in which? The commit message does not explain that, nor does it link to a Phabricator ticket. Not that this is an issue in general. We are very thankful for patches like this in almost all cases. But when questions arise, we have to ask them. Basically, how did you found these places in the code and why do you think they need to change? I would like to understand.

Patch-set: 1