Update patch set 1
Patch Set 1:

The reason for the patch is the FIXME.

All new Language::factory() are in the tests. It does not introduce a new Language::factory() in the productive code. It just removes one.

In the current productive code a Language object gets converted to a language code string and converted back to a Language object to get the direction of the language. This change keeps the Language object and uses it directly.

Patch-set: 1
Label: Verified=0