Update patch set 5
Patch Set 5:

(2 comments)

I don't see why we need to do magic dispatching based on inspecting the contents of the statement list. We already know what kind of entity we want to show. That's what EntityViewFactory does. 

In my mind, StatementSectionsView should know a StatementGrouper. EntityViewFactory creates the right grouper using a StatementGrouperFactory, constructs a StatementSectionsView based on that, and injects that into the EntityView.

That seems much more straight forward to me. I'd expect that to be less code too, since we don't need the dispatch logic.

Patch-set: 5