Update patch set 12
Patch Set 12:

@daniel i've manually tested, but would be cool if someone else sets up cirrus and tries this. (should be easy with vagrant)

nothing has to be done specifically at deploy time. I tried what happens (nothing terrible or problematic) when an item is saved, but the mapping is not updated yet. at soonest convenience, after deploy, it would be good to update the mapping for wikidata and then reindex. (like we did for coordinates)

Patch-set: 12