Update patch set 5
Patch Set 5:

(1 comment)

EntityViewFactory is in the view component, not in repo. What you suggest is not possible without making view depend on repo, or create a "StatementGroupFactory" interface, or add a type parameter to the StatementGrouper interface. I find all this weird. It's also not necessary, as this patch shows.

Patch-set: 5