Update patch set 12
Patch Set 12:

Now we are both arguing with technical dept. :-( I do not know how anything in that class can be more than a single line instantiating a new object. I wonder if this class will ever be used in a second place. I assume this will never happen. So what's the benefit of introducing it *now*? Why not introduce it *later* when the array is big enough?

Patch-set: 12