Update patch set 12
Patch Set 12:

@thiemo ... ok, you are asking about WikibaseFieldDefinitions and why not put this into the Cirrus hooks handler?

ideally, i'd like most of this code to not be in wikibase (maybe in a component) and only depend on data model and elastica.  then it is more reusable for tools and other stuff. (i already experimented with indexing dumps in elastic with only those dependencies and it's helpful)

Patch-set: 12