Fix broken references-heading CSS
The selector was broken. And it should use margin, because margins do
not stack. This is how it should be here. JS add an <a> with the same
margin-top: 0.5em. The additional margin here does have the effect that
our entity pages look more similar in no-JS and JS mode.

Change-Id: I5ad7bd10f20cd7cbf95a27ac1250cce3f1c5cd4c