Add button icon when missing in the HTML template
Yes, this is a hack. But it's completely opaque (or transparent,
depends on how you look at it). If everything is fine, the <span> is
there and the "length" check will fail and do nothing. If the <span>
is missing because the template delivered by ResourceLoader is
outdated (and yes, this is relevant) the span is added.

Bug: T87757
Change-Id: If192189f07f7a105172faf50bcb82df307bfb2f7