Update patch set 5
Patch Set 5:

I still wonder what we are discussing here. I still have the feeling we are targeting different goals instead of working together. This still creates friction and frustration.

I agree with Adrian. What the StatementGrouperFactory does is specific for the repo and does not belong to the view.

Patch-set: 5