Update patch set 12
Patch Set 12:

I'm mainly asking for WikibaseFieldDefinitions. It contains nothing but configuration. It should be configuration and not a class in that namespace.

I guess you can argue that all the ...Field classes are part of this configuration and therefore affected by the same argument. I'm not sure if this is true. I currently do not have a problem with these classes.

Patch-set: 12