Update patch set 1
Patch Set 1:

This does not block anything else. It's straightforward fail-safe code that will do nothing if everything is right, and will avoid incomplete buttons in rare edge cases. What's wrong with that? What's the reason for the -1? What's the argument you "do not want to argue about"? I'm not sure on what assumptions your -1 is build. As I said this is not about parser output. Nothing Aude did is related to this patch, as far as I know.

Patch-set: 1