Update patch set 5
Patch Set 5:

@Thiemo I think our goals are the same (simple and flexible code), but we have different opinions on what exactly that means and how to get there. And differences in opinion can and should be discussed in code review. But they should not lead to a stalemate.

To me, grouping statements is functionality of the view. It doesn't have anything to do what the repos does, except for viewing. Also, my suggestion to inject the grouper into the StatementListView instead of the EntityView is independant of the question whether we need a DispatchingEntityTypeStatementGrouper. I simply don't see why we need it. We can just create the right grouper for each type of entity directly in the EntityViewFactory and be done. Less code, less magic - or am I missing something?

Anyway, I'm not totally opposed to this solution, I just think it's more complicated than necessary.

Patch-set: 5