Update patch set 5
Patch Set 5:

Since a patch says more than a thousand words, here is what I mean: Icaeeec

That removes 150 lines of code, without changing functionality, or any refactoring even.

The dispatching thing isn't needed here, since the info the dispatch is based on is already known at the time it is created. So a factory is more suitable, and simpler.

Patch-set: 5