Update patch set 5
Patch Set 5:

> grouping statements is functionality of the view.

It is. The view gets a grouper and does the grouping.

> my suggestion to inject the grouper into the StatementListView instead of the EntityView

I will do that in a later patch, see Ic458864.

> We can just create the right grouper for each type of entity directly in the EntityViewFactory

No, we can't. Creating the "right" groupers is repo knowledge. This can not be done by the view. This would make view depend on repo, maybe not on a class but on knowledge about repo.

Patch-set: 5