Update patch set 5
Patch Set 5:

@thiemo: you said:

> No, we can't. Creating the "right" groupers is repo knowledge. This can not be done by the view. This would make view depend on repo, maybe not on a class but on knowledge about repo.

Which knowledge? You mean the structure of the configuration array? Why should that be bound to the repo? It's configuration of the view, specifically.

The place that has all the knowledge about how to construct a view for a given entity type is EntityViewFactory - which is in the view component. 

I don't understand what dependency you are trying to avoid.

Patch-set: 5
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0