Update patch set 1
Patch Set 1:

Ok, here is a way to look at this: Imagine we want to group statements based on the mood of the user. This means we need a User object to create a StatementGrouper. With this patch this creates the situation where we need to pass this object to view, making view depend on MediaWiki core.

I had the impression this is something we want to avoid.

Patch-set: 1